Split DecUT_MotionCompensation to a shareable header
This allows adding a version of the same test for the encoder.
